Add ability to change Hunspell library in runtime
